build.gradle actualizado y sistema de paquetes Clean Arquitecture
