+++ app/src/main/java/com/example/servivelog/ui/view/FragmentLogin.kt
@@ -1,0 +1,26 @@
+package com.example.servivelog.ui.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.servivelog.databinding.FragmentLoginBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class FragmentLogin : Fragment() {
+
+    private lateinit var loginBinding: FragmentLoginBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        loginBinding = FragmentLoginBinding.inflate(layoutInflater)
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        return loginBinding.root
+    }
+}
