+++ app/src/main/java/com/example/servivelog/ui/view/MainActivity.kt
@@ -1,8 +1,11 @@
-package com.example.servivelog
+package com.example.servivelog.ui.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.servivelog.R
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
