+++ app/src/main/res/layout/fragment_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FragmentLogin">
+    tools:context=".ui.view.FragmentLogin">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -49,6 +49,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:hint="Ingrese su usuario"
+                tools:text="Moreno32"
                 android:inputType="text" />
         </com.google.android.material.textfield.TextInputLayout>
