+++ app/src/main/res/navigation/nav_grahp.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/fragmentLogin"
-        android:name="com.example.servivelog.FragmentLogin"
+        android:name="com.example.servivelog.ui.view.FragmentLogin"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" />
 </navigation>
